+++ packages/impulse-bridge/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gravito/impulse-bridge
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies [1e7f7d6]
+  - @gravito/impulse@1.1.0
+
 ## 1.0.2
 
 ### Patch Changes